+++ semgrep-core/tainting/tainting_generic.mli
@@ -1,0 +1,3 @@
+val check: 
+  Tainting_rule.rules ->  Common.filename -> Ast_generic.program ->
+  Match_result.t list